refactor(login): remove old login component and integrate new authentication logic
