+++ apps/sansa/src/app/login/page.tsx
@@ -1,5 +1,3 @@
-// app/login
-
 'use client'
 
 import { useState, FormEvent } from 'react'
